+++ src/utils/localstorage.ts
@@ -39,6 +39,30 @@ export function getLocalstorageWithSchema<Schema extends z.ZodTypeAny>(
   return out!;
 }
 
+export function getNullableLocalstorageWithSchema<Schema extends z.ZodTypeAny>(
+  key: string,
+  schema: Schema,
+): z.infer<Schema> | undefined {
+  let value: unknown;
+  const valueRaw = localStorage.getItem(key);
+  if (valueRaw === null) {
+    return undefined;
+  }
+
+  try {
+    value = JSON.parse(valueRaw);
+  } catch (err) {
+    throw new ErrorInvalidJson(key);
+  }
+
+  const results = schema.safeParse(value);
+  if (!results.success) {
+    throw results.error;
+  }
+  const out = results.data as z.infer<Schema>;
+  return out!;
+}
+
 export function setLocalstorageWithSchema<
   Schema extends z.ZodTypeAny,
   DataType extends z.infer<Schema> | undefined,